# Optimize scanning routines in the project manager