@KoBeWi 2023-12-20
The only problem with scanning big folders is that Project Manager does not show any progres whatsoever (#5953). User might pick something very big for scan and ends up with unresponsive window.
Though scanning times are fine from my testing. 100+ are founds within seconds in a dev build, which is acceptable. But I did not test scanning whole drive 🙃