@YuriSizov 2024-01-25
Cherry-picked for 4.2.2.